refactor[notask]: move multi-gpu-mobile util to server/utils, clarify comment
