+++ packages/sdk/test/unit/multi-gpu-mobile.test.ts
@@ -1,6 +1,6 @@
 // @ts-expect-error brittle has no type declarations
 import test from "brittle";
-import { stripMultiGpuKeys, MULTI_GPU_KEYS } from "@/server/bare/plugins/utils/multi-gpu-mobile";
+import { stripMultiGpuKeys, MULTI_GPU_KEYS } from "@/server/utils/multi-gpu-mobile";
 
 test("stripMultiGpuKeys: removes all three multi-GPU keys when present", (t) => {
   const config: Record<string, unknown> = {
